handle exceptions for students

+++ src/main/java/com/jj/Gradebook/service/student/StudentServiceImpl.java
@@ -2,6 +2,8 @@ package com.jj.Gradebook.service.student;
 
 import com.jj.Gradebook.dao.StudentRepository;
 import com.jj.Gradebook.entity.Student;
+import com.jj.Gradebook.exceptions.EntityAlreadyExistException;
+import com.jj.Gradebook.exceptions.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -21,28 +23,50 @@ public class StudentServiceImpl implements StudentService{
     }
 
     @Override
-    public Student findById(int id) {
+    public Student findById(int id) throws EntityNotFoundException {
         Optional<Student> result = studentRepository.findById(id);
 
-        Student student = null;
         if(result.isPresent()){
-            student = result.get();
+            return result.get();
         }
-        else{ //TODO: FIND BETTER APPROACH
-            throw new RuntimeException("No student with id - " + id);
+        else {
+            throw new EntityNotFoundException("No student with id - " + id);
         }
-        return student;
+
+    }
+
+    @Override
+    public Student findByPesel(String pesel) throws EntityNotFoundException {
+        Optional<Student> result = studentRepository.findStudentByUserPesel(pesel);
+        if (result.isPresent()){
+            return result.get();
+        }
+        else {
+            throw new EntityNotFoundException("No student with pesel - " + pesel);
+        }
     }
 
     @Override
     @Transactional
-    public Student save(Student student) {
-        return studentRepository.save(student);
+    public Student save(Student student) throws EntityAlreadyExistException {
+        Optional<Student> existingStudent = studentRepository.findStudentByUserPesel(student.getUser().getPesel());
+        if (existingStudent.isEmpty()){
+            return studentRepository.save(student);
+        }
+        else {
+            throw new EntityAlreadyExistException("Student already exists!");
+        }
     }
 
     @Override
     @Transactional
-    public void deleteById(int id) {
-        studentRepository.deleteById(id);
+    public void deleteById(int id) throws EntityNotFoundException {
+        Optional<Student> existingStudent = studentRepository.findById(id);
+        if (existingStudent.isPresent()){
+            studentRepository.deleteById(id);
+        }
+        else {
+            throw new EntityNotFoundException("No student with id - " + id);
+        }
     }
 }
